Half-way through investigating svg weirdness
It seems the SVG is being rendered as a square. If you resize the
window, when the width of the SVG becomes greater than the fill
portion it has available it starts to push the size of the whole
thing! Play a little bit with resizing the window.

Also, playing with the values of `flex-grow` on the topRow seems
to force the SVG to shrink (when raising the size of its sibling
on the left. Try it!